finialize font script
