+++ tools/outline-svg/main.mjs
@@ -1,31 +1,68 @@
 import { promises as fs } from 'fs';
-import path from 'path'
 import outlineStroke from 'svg-outline-stroke';
 import { parse, stringify } from 'svgson'; // eslint-disable-line import/no-extraneous-dependencies
 import getArgumentOptions from 'minimist';
-import Tinypool from 'tinypool'
+import path from 'path';
 
-
-const inputDir = path.join(process.cwd(), '../../icons')
+const inputDir = path.join(process.cwd(), '../../icons');
 const cliArguments = getArgumentOptions(process.argv.slice(2));
 const { outputDir = 'outlined' } = cliArguments;
-const targetDir = path.join(process.cwd(), '../../', outputDir)
+const targetDir = path.join(process.cwd(), '../../', outputDir);
+
+function transformForward(node) {
+  if (node.name === 'svg') {
+    return {
+      ...node,
+      attributes: {
+        ...node.attributes,
+        width: 960,
+        height: 960,
+      },
+    };
+  }
+  return node;
+}
+
+function transformBackwards(node) {
+  if (node.name === 'svg') {
+    const { width, height, ...attributes } = node.attributes;
+    return {
+      ...node,
+      attributes,
+    };
+  }
+  return node;
+}
 
 async function init() {
-  const pool = new Tinypool({
-    filename: new URL('./outlineSvg.mjs', import.meta.url).href,
-  })
-
   console.time('icon outliner');
   try {
-    await fs.mkdir(targetDir);
-  } catch (err) {}
+    try {
+      await fs.mkdir(targetDir);
+    } catch (error) {} // eslint-disable-line no-empty
 
-  try {
     const icons = await fs.readdir(inputDir);
+    const parsedIconNodes = await Promise.all(
+      icons.map(async (file) => {
+        const inputFilePath = path.resolve(process.cwd(), inputDir, file);
+        const iconContent = await fs.readFile(inputFilePath);
+        const iconNode = await parse(iconContent.toString(), {
+          transformNode: transformForward,
+        });
+        return [file, iconNode];
+      }),
+    );
 
     await Promise.all(
-      icons.map(file => pool.run({ file, inputDir, targetDir })),
+      parsedIconNodes.map(async ([file, iconNode]) => {
+        const outlined = await outlineStroke(stringify(iconNode));
+        const outlinedWithoutAttrs = await parse(outlined, {
+          transformNode: transformBackwards,
+        });
+
+        const filePath = path.join(targetDir, file);
+        await fs.writeFile(filePath, stringify(outlinedWithoutAttrs));
+      }),
     );
 
     console.timeEnd('icon outliner');
