+++ tools/outline-svg/package.json
@@ -4,8 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "node ./main.mjs",
-    "start:old": "node ./outlineSvg.old.mjs"
+    "start": "node ./main.mjs"
   },
   "keywords": [],
   "author": "",
@@ -13,7 +12,6 @@
   "dependencies": {
     "minimist": "^1.2.6",
     "svg-outline-stroke": "^1.3.1",
-    "svgson": "^5.2.1",
-    "tinypool": "^0.3.0"
+    "svgson": "^5.2.1"
   }
 }
